null dest

+++ AmpRageRepo/Models/Waypoint.cs
@@ -12,7 +12,7 @@ namespace AmpRageRepo.Models
         public double Longitude { get; set; }
         public int? TimeToCharge { get; set; }
         public string CoordString { get; set; }
-        public double? Emissions { get; set; }
+        public decimal? Emissions { get; set; }
 
         public ChargingStationRootObject ChargingStation { get; set; }
     }
